+++ src/components/Main/Description.test.jsx
@@ -1,5 +1,5 @@
 import Description from './Description'
-import { render, screen } from '../utils/test-utils'
+import { render, screen } from '../../utils/test-utils'
 
 describe('text in the document', () => {
   it('title is visible', () => {
