+++ src/components/UI/Modal/SettingsModal.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import { useEffect, useState } from 'react'
 import Modal from './Modal'
 import SettingsModalRow from './SettingsModalRow'
-import TextInput from '../../TextInput'
+import TextInput from '../Inputs/TextInput'
 import Toggle from '../Toggle/Toggle'
 
 SettingsModal.propTypes = {
